+++ lib/inview_notifier_list.dart
@@ -1,3 +1,3 @@
 export 'src/inview_notifier_list.dart';
 export 'src/inview_state.dart';
-export 'src/inview_notifier.dart' hide InViewNotifier;
+export 'src/inview_notifier.dart';